Rename components - #4683
