+++ ui/app/src/components/PublishingDialog/index.ts
@@ -14,6 +14,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { default } from './PublishingStatus';
+export { default } from './PublishingDialog';
 
-export * from './PublishingStatus';
+export * from './PublishingDialog';
